requirements.txt
Django==1.8
django-appconf==1.0.2
django-compressor==1.6
django-debug-toolbar==1.2.2
django-filter==0.12.0
djangorestframework==3.3.2

Markdown==2.6.5
psycopg2==2.5.4
requests==2.11.1
sqlparse==0.2.1

# django-pgjson library that add a bridge for PostgreSQL json field to Django.
# For API support http://djangonauts.github.io/django-pgjson/
django-pgjson==0.3.1
